Fixed bugs with permissions; gui updating

+++ plugin/src/main/kotlin/ru/astrainteractive/astrashop/gui/buy/BuyGUI.kt
@@ -53,7 +53,7 @@ class BuyGUI(shopConfig: ShopConfig, item: ShopConfig.ShopItem, override val pla
     private fun setActionButton(type: BuyType, i: Int, state: BuyState.Loaded) {
         val amount = 2.0.pow(i).toInt()
         if (type == BuyType.BUY && state.item.stock != -1 && state.item.stock < amount) return
-        val totalPrice = (amount * state.item.price).toInt().coerceAtLeast(0)
+        val totalPrice = (amount * state.item.price).coerceAtLeast(0.0)
 
         val title = when (type) {
             BuyType.BUY -> translation.buttonBuyAmount(amount)
